% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poster.R
\name{covMatrix}
\alias{covMatrix}
\title{Title}
\usage{
covMatrix(X1, X2 = NULL, theta, correlation = "gauss")
}
\arguments{
\item{X1}{}

\item{X2}{}

\item{theta}{}

\item{correlation}{}
}
\description{
Title
}
