% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poster.R
\name{create_entropies}
\alias{create_entropies}
\title{Title}
\usage{
create_entropies(gpi, xdiscrete, theta, X, y)
}
\arguments{
\item{gpi}{}

\item{xdiscrete}{}

\item{theta}{}

\item{X}{}

\item{y}{}
}
\description{
Title
}
