% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poster.R
\name{GPupdate}
\alias{GPupdate}
\title{Title}
\usage{
GPupdate(
  X,
  X.cur,
  y.cur = NULL,
  X.add,
  y.add = NULL,
  theta,
  mean.cur,
  Sigma,
  sigma = NULL,
  mean.add = NULL,
  correlation = "gauss"
)
}
\arguments{
\item{X}{}

\item{X.cur}{}

\item{y.cur}{}

\item{X.add}{}

\item{y.add}{}

\item{theta}{}

\item{mean.cur}{}

\item{Sigma}{}

\item{sigma}{}

\item{mean.add}{}

\item{correlation}{}
}
\description{
Title
}
